pyproject.toml
[project]
name = "learnings"
version = "0.1.0"
description = "Repository for learnings"
authors = [
    { name = "Sarvagya Meel", email = "sarvagyameel2@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "python-dotenv >=1.1.0,<2.0.0",
    "requests >=2.32.3,<3.0.0",
    "autogen >=0.9,<0.10",
    "ag2[openai] (>=0.9,<0.10)",
    "openai (>=1.77.0,<2.0.0)",
    "google-adk (>=1.0.0,<2.0.0)",
    "litellm (>=1.71.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false